dom/webauthn/tests/test_webauthn_webdriver_virtual_authenticator.html

<!DOCTYPE html>
<meta charset=utf-8>
<head>
  <title>Tests for WebDriver Virtual Authenticator Extension for W3C Web Authentication</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="u2futil.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

  <h1>Tests for WebDriver Virtual Authenticator Extension for W3C Web Authentication</h1>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1460986">Mozilla Bug 1460986</a>

  <script class="testbody" type="text/javascript">
    "use strict";

    function arrivingHereIsGood(aResult) {
      ok(true, "Good result! Received a: " + aResult);
    }

    function arrivingHereIsBad(aResult) {
      ok(false, "Bad result! Received a: " + aResult);
    }

    function expectNotAllowedError(aResult) {
      ok(aResult == "NotAllowedError", "Expecting a NotAllowedError, got " + aResult);
    }

    function getAssertion(id) {
        let chall = new Uint8Array(16);
        crypto.getRandomValues(chall);

        let options = {
          challenge: chall,
          allowCredentials: [ { type: "public-key", id } ],
        };

        return navigator.credentials.get({publicKey: options});
    }

    add_task(async function test_add_and_remove_credential() {
      let authenticatorId = await addVirtualAuthenticator();
      let credIdB64 = await addCredential(authenticatorId, document.domain);
      let credId = base64ToBytesUrlSafe(credIdB64);

      await getAssertion(credId)
          .then(arrivingHereIsGood)
          .catch(arrivingHereIsBad);

      await removeCredential(authenticatorId, credIdB64);
      await getAssertion(credId)
          .then(arrivingHereIsBad)
          .catch(e => expectNotAllowedError(e.name));
    });
  </script>

</body>
</html>
